Fix state change after unmount in StartQuizButton
The firebase update promise can resolve after the parent component has
unmounted the StartQuizButton. Hopefully adding a ref to act as an
unmount sentinal will prevent that (and the resulting error in the
console). It's hard to test the race condition, but it seems to work.

It would be better to cancel the update promise somehow, but that doesn't
seem possible