Replace useEffect with useRef in useStartTime
Previously, useStartTime internally used useEffect, effectively to test
if the reset key had changed. This worked, but meant the result was only
updated after the DOM was rendered.

By using useRef and manually checking if the key changes, the result of
useStartTime is updated synchronously.

The net result is that timers reset instantly when the owner clicks
'next', rather than after a short pause.