Autoscroll scroll listener add/remove improvements
updateAutoscrollDebounced now makes use of a ref rather than a state
value to determine if autoscroll is enabled. It therefore does not
change across renders, making adding/removing it as an event listener
more reliable.

Additionally the useEffect cleanup action was not necessary, because
setContainerRef is called with null when the component unmounts.